refactor: make packages for model and llm_integration
