+++ src/llm_integration/data_extraction.py
@@ -1,7 +1,7 @@
 from typing import List, Dict
-from app_model import AppModel
+from model.app_model import AppModel
 import json
-from llm_service import LLM, GPT4o
+from llm_integration.llm_service import LLM, GPT4o
 from error_handling import ValidationException
